refactor: corrects variable naming after review
