+++ src/components/AssetMatching/components/ActionsToolbox/ActionsToolbox.tsx
@@ -6,20 +6,20 @@ import { flex } from 'theme/functions';
 
 interface Props {
   matchingActions?: MatchingAction[];
-  actionButtonFill?: boolean;
+  isButtonFilled?: boolean;
 }
 
-const ActionsToolbox: FC<Props> = ({ matchingActions = [], actionButtonFill = false }) => {
+const ActionsToolbox: FC<Props> = ({ matchingActions = [], isButtonFilled = false }) => {
   const primaryActions = matchingActions.slice(0, 2);
   const secondaryActions = matchingActions.slice(2, matchingActions.length);
 
   return (
     <div css={flex}>
       {hasActions(primaryActions) && (
-        <PrimaryActions actionButtonFill={actionButtonFill} primaryActions={primaryActions} />
+        <PrimaryActions isButtonFilled={isButtonFilled} primaryActions={primaryActions} />
       )}
       {hasActions(secondaryActions) && (
-        <SecondaryActions actionButtonFill={actionButtonFill} secondaryActions={secondaryActions} />
+        <SecondaryActions isButtonFilled={isButtonFilled} secondaryActions={secondaryActions} />
       )}
     </div>
   );
