+++ src/components/AssetMatching/components/utils.tsx
@@ -35,12 +35,12 @@ export const useCategoryItemActions = (item: string, matchedCategoryItems?: stri
   };
 };
 
-export const createActionButton = (filled = false) => (action: MatchingAction) => (
+export const createActionButton = (isButtonFilled = false) => (action: MatchingAction) => (
   <Button
     type={'primary'}
     color={'neutralBlack-700'}
     iconLeft={<Icon color={'inherit'} name={action.icon} />}
-    filled={filled}
+    filled={isButtonFilled}
     onClick={action?.onClick}
   >
     {action.text}
@@ -50,11 +50,11 @@ export const createActionButton = (filled = false) => (action: MatchingAction) =
 export const useMatchingActions = (
   actions: MatchingAction[],
   enhanceWithWrapperElement: (actionButton: JSX.Element, index: number) => JSX.Element,
-  actionButtonFill = false
+  isButtonFilled = false
 ) => {
   const actionItems = useMemo(
-    () => actions.map(createActionButton(actionButtonFill)).map(enhanceWithWrapperElement),
-    [actions, enhanceWithWrapperElement, actionButtonFill]
+    () => actions.map(createActionButton(isButtonFilled)).map(enhanceWithWrapperElement),
+    [actions, enhanceWithWrapperElement, isButtonFilled]
   );
 
   return { actionItems };
